requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
drf-writable-nested==0.6.2
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
